Initial Tests
Ok, let's start out with the basic fact that this **is not** garbage. The client
code is fine. I've augmented it some. Neither of us are experience JS devs,
especially when it comes to ES6/7 stuff. I found a bunch of gotchas once I started
running teses.

I kept the garbage name though -- it's funny.

** Stuff I Changed **

** lib/index.js **

This doesn't need to instantiate anything: actually, all you need to do (well,
want more than need) is to export the classes or functions or whatever you want
to be top level from the other modules. It's very much like the package init
file in Python.

** lib/client.js **

We should keep consistent in import styles, and I think the preferred way to
do it is the `import 'blah'` or `import { whatever } from 'blah'`. Also, we want
the whole scope of node-fetch -- I only found this after my tests kept bombing
and I did some debugging.

We don't want to export `BaseClient` since we're using it as an abstract
base; I removed that directive.

The Fetch API returns a promise, which is inherently asynchronous. In ES6 or
5 or whatever they introduced the async/await pattern. For functions that return
promises, we should label them as async and await their response.

Fetch also has a quirk where you need two `then`s. The first is to check if
the response is good and parse the body, the second is to handle the body. The
error block is only for hard 4xx and 5xx. It's stupid.

In `describeDataSets`, we need to return the response from `sendRequest`. JS
doesn't do transparent fall throughs. Again, only exposed after running the tests.

** package.json **

I got rid of the `build` script -- we don't need it right now and I think it may
be messing with stuff.

** Stuff I Added **

A few new dependencies and fixtures for testing.

** test/client.test.js **

I removed the example test module (it sucked) and wrote some tests for the client
as it is now.

I've opted to use `chai/expect`/BDD style tests because:

1. This is a greenfield library and it may help to write the tests before we
   write the code (just to keep shit sane)
2. None of us are _good_ at JS, so again it's a way to enforce we're doing things
   the right way. If we make a contract in the test, we fulfill it with the code.